Update DataRequest.java
Changed method names to insertRecord and updateRecord, that way we can just overload; I also changed each to return a boolean value signifying whether the operation requested was successful.
Also, please see the comments I added about including the table build in this class -- not sure that's a good idea; The way it's coded, it looks like every time we establish a connection to the database (i.e. every time the program starts), we'll be rebuilding the tables. Maybe this is part of a DB setup that needs to happen every time?
Lastly, updated the ID's to be Integers instead of Longs in the method signatures, per our chat last night -- why are we passing in both the ID and the object though? If we're passing in the object, you can just get the ID from the object, no?